rename bifacial_radiance.py to main.py to avoid import collisions
