+++ bifacial_radiance/__init__.py
@@ -1,4 +1,4 @@
 from __future__ import absolute_import 
-from bifacial_radiance.bifacial_radiance import AnalysisObj, GroundObj, MetObj, RadianceObj, SceneObj
+from bifacial_radiance.main import AnalysisObj, GroundObj, MetObj, RadianceObj, SceneObj
 from bifacial_radiance.readepw import readepw
 from bifacial_radiance import load
